+++ exec/source.go
@@ -3,7 +3,6 @@ package exec
 import (
 	u "github.com/araddon/gou"
 
-	"github.com/araddon/qlbridge/datasource"
 	"github.com/araddon/qlbridge/plan"
 	"github.com/araddon/qlbridge/schema"
 )
@@ -30,12 +29,12 @@ var (
 type Source struct {
 	*TaskBase
 	sp      *plan.SourcePlan
-	scanner datasource.Scanner
+	scanner schema.Scanner
 	JoinKey KeyEvaluator
 }
 
 // A scanner to read from data source
-func NewSource(sp *plan.SourcePlan, scanner datasource.Scanner) *Source {
+func NewSource(sp *plan.SourcePlan, scanner schema.Scanner) *Source {
 	s := &Source{
 		TaskBase: NewTaskBase(sp.Ctx, "Source"),
 		scanner:  scanner,
@@ -45,7 +44,7 @@ func NewSource(sp *plan.SourcePlan, scanner datasource.Scanner) *Source {
 }
 
 // A scanner to read from sub-query data source (join, sub-query)
-func NewSourceJoin(sp *plan.SourcePlan, scanner datasource.Scanner) *Source {
+func NewSourceJoin(sp *plan.SourcePlan, scanner schema.Scanner) *Source {
 	s := &Source{
 		TaskBase: NewTaskBase(sp.Ctx, "SourceJoin"),
 		scanner:  scanner,